update: conn wrap logic
