+++ internal/socket/tcpconn.go
@@ -2,17 +2,80 @@ package socket
 
 import (
 	"fmt"
+	"io"
 	"net"
+	"os"
 	"sync"
+	"time"
+
+	"github.com/gaukas/water/internal/log"
 )
 
-func TCPConnPair(address ...string) (c1, c2 net.Conn, err error) {
-	var tcpAddr string = ":0"
-	if len(address) > 0 && address[0] != "" {
-		tcpAddr = address[0]
+// TCPConnFileWrap wraps an object into a *os.File from an
+// underlying net.TCPConn. The object must implement io.Reader
+// and/or io.Writer.
+//
+// If the object implements io.Reader, upon completing copying
+// the object to the returned *os.File, the callback functions
+// will be called.
+//
+// It is caller's responsibility to close the returned *os.File.
+func TCPConnFileWrap(obj any, callbacks ...func()) (*os.File, error) {
+	// get a pair of connected UnixConn
+	tcpConn, reverseTCPConn, err := TCPConnPair()
+	if err != nil && (tcpConn == nil || reverseTCPConn == nil) {
+		return nil, err
 	}
 
-	l, err := net.Listen("tcp", tcpAddr) // skipcq: GSC-G102
+	tcpConnFile, err := tcpConn.File()
+	if err != nil {
+		return nil, err
+	}
+
+	// if the object implements io.Reader: read from the object and write to the reverseUnixConn
+	if reader, ok := obj.(io.Reader); ok {
+		go func() {
+			_, _ = io.Copy(reverseTCPConn, reader)
+			// when the src is closed, we will close the dst
+			time.Sleep(1 * time.Millisecond)
+			log.Debugf("closing reverseTCPConn and tcpConn")
+			for _, f := range callbacks {
+				f()
+			}
+			_ = reverseTCPConn.Close()
+			_ = tcpConn.Close()
+		}()
+	}
+
+	// if the object implements io.Writer: read from the reverseUnixConn and write to the object
+	if writer, ok := obj.(io.Writer); ok {
+		go func() {
+			_, _ = io.Copy(writer, reverseTCPConn)
+			// when the src is closed, we will close the dst
+			if closer, ok := obj.(io.Closer); ok {
+				time.Sleep(1 * time.Millisecond)
+				log.Debugf("closing obj")
+				_ = closer.Close()
+			}
+		}()
+	}
+
+	return tcpConnFile, nil
+}
+
+// TCPConnPair returns a pair of connected net.TCPConn.
+func TCPConnPair(address ...string) (c1, c2 *net.TCPConn, err error) {
+	var addr string = ":0"
+	if len(address) > 0 && address[0] != "" {
+		addr = address[0]
+	}
+
+	tcpAddr, err := net.ResolveTCPAddr("tcp", addr)
+	if err != nil {
+		return nil, nil, fmt.Errorf("net.ResolveTCPAddr returned error: %w", err)
+	}
+
+	l, err := net.ListenTCP("tcp", tcpAddr) // skipcq: GSC-G102
 	if err != nil {
 		return nil, nil, fmt.Errorf("net.Listen returned error: %w", err)
 	}
@@ -22,10 +85,10 @@ func TCPConnPair(address ...string) (c1, c2 net.Conn, err error) {
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		c2, goroutineErr = l.Accept()
+		c2, goroutineErr = l.AcceptTCP()
 	}()
 
-	c1, err = net.Dial("tcp", l.Addr().String())
+	c1, err = net.DialTCP("tcp", nil, l.Addr().(*net.TCPAddr))
 	if err != nil {
 		return nil, nil, fmt.Errorf("net.Dial returned error: %w", err)
 	}
